pyproject.toml
[project]
name = "lyricsviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.18"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20250625",
]
